Best Laid Plans
I threw in some skelleton methods for us to fill out in the hero class
for later. I made a jumping method, it doesn't actually handle the bulk
work of jumping but it tells the hero that he should be jumping in the
move method and handles the decrementing of the jumpCounter.

Some stuff to work on tonight/soon
*Set facing direction
*Aiming
*Drawing Mouse on screen
*Crouching
*Crawling
*Attacking
*Character movement - Camera follows character or static???
*Jumping
*Gravity
*Moving Left Right